Balls are moving now, it works not so good yet but bugs will be fixed soon. Problem with double-type direction was solved.

+++ src/com/graphics/Direction.java
@@ -2,8 +2,8 @@ package com.graphics;
 
 public class Direction {
     // Private
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     // Public
     public Direction() {
@@ -11,26 +11,26 @@ public class Direction {
         y = 0;
     }
 
-    public Direction(int x, int y) {
+    public Direction(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     // Getters
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
     // Setters
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
